backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.4.2
python-docx==1.0.1
PyMuPDF==1.23.8
huggingface_hub==0.16.4
sentence-transformers==2.2.2
faiss-cpu==1.7.4
openai==1.3.0
python-jose==3.3.0
passlib==1.7.4
python-dotenv==1.0.0
chromadb==0.4.15
langchain==0.0.350 